move release parsing to files for each version
